requirements.txt
Django==1.6.4
MySQL-python==1.2.5
Pillow==2.4.0
boto==2.27.0
django-secretballot==0.2.3
django-storages==1.1.8
wsgiref==0.1.2
